Airflow/AirFlow_DEPI/airflow_depi/requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages

pyodbc
apache-airflow-providers-microsoft-mssql
apache-airflow-providers-odbc
